operand_stack become 64bit

+++ sperl_vm.h
@@ -5,7 +5,7 @@
 
 // Field information
 struct SPerl_vm {
-  int32_t* operand_stack;
+  int64_t* operand_stack;
   int32_t* call_stack;
   int32_t call_stack_capacity;
   int32_t operand_stack_capacity;
